Report size in bytes for all operating systems
Currently, we report back a string containing whatever the operating
system decides is a human-readable size, like "7.8G", however this is
not an ideal format to perform operations on the resulting sizes.

- OS X

The size in bytes is reported by `diskutil info` in parenthesis at the
right of the human-readable size we're currently showing. We use `perl`
and a regular expression to capture the result.

I tried to implemented in `awk` in order to not make use of yet another
tool, however capturing regular expression groups with `awk` proved to
be a pain, or it relied on GNU awk specific features.

- Linux

`lsblk` contains an option `-b` to display the size in bytes rather than
in a human readable format.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>

- Windows

The size property we were getting from the drive object was already in
bytes, but we were manually transforming to a human-readable format
before for consistency reasons.