Print stderr script output to stderr if exit code is non zero
Platform scripts might output warnings to `stderr` that we don't want to
take into consideration if the script exits with an exit code 0.

If the scripts exits with a non-zero code, `.execFile()` yields an error
whose message includes the `stderr` output, so we don't need to do
anything ourselves.

See: https://github.com/resin-io/etcher/issues/1034
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>