Implement drivelist.get()
This function allows the user to get information about a single drive.
Currently, this function runs the whole drive detection script and
filters the output, however in the future we might update the scripts to
take a drive argument, and only print information about that one, which
should be way faster.

See: https://github.com/resin-io/etcher/issues/750
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>