fix(windows): all drives are interpreted as read-only
This was an accidental subtle logical flaw in the PR where we check both
the disk and the volume writable state.

We only want to consider the volumes writable state if the disk is
read-only, and the volumes have a file-system.

Change-Type: patch
See: https://github.com/resin-io-modules/drivelist/pull/183
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>