refactor: reduce debug logging if unlink was successful
As a way to reduce the amount of debug logging this module produces, we
only emit an unlink debug message if the unlink operation was succesful.

Also, as a better way to correlate unlink issues with a certain
drivelist call, we only report back the result of this module after the
unlink operation ended (sucessfully or not).

Change-Type: patch
See: https://github.com/resin-io/etcher/pull/1600
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>