fix: workaround noexec by executing scripts with bash
The temporary or `XDG_RUNTIME_DIR` directory might be mounted with the
`noexec` option, which means we might not have execution permissions to
run the scripts we just extracted.

As an easy workaround, we can call `bash` passing it the script as an
argument rather than executing the script directly.

Change-Type: patch
Changelog-Entry: Correctly execute macOS and GNU/Linux scripts inside directories mounted as `noexec`.
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>