feat: Rewrite native bindings to use IOCTL
This completely refactors the native bindings on Windows
to avoid COM/WMI etc. and solely rely on DeviceIoControl.
This also introduces some new properties to the device,
and fixes a range of errors / bugs.

Further changes / fixes:

- attached virtual hard drives don't get labeled as system drives
anymore
- errors with one device don't stop enumeration of other devices (and
don't cause errors to be thrown)
- info about bus / protocol (SCSI, SATA, USB, UAS, etc.)
- isReadonly is now only true if the device itself is readonly (i.e. the
switch on the sd-card adapter is flicked), not when a fs or volume is
mounted or marked as readonly
- added SD-card detection
- added detection of virtual storage devices
- BitLocker volumes no longer cause trouble, and devices containing one
will be enumerated as any other storage device

Change-Type: major